+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- `QDateTime::from_string` to parse `QDateTime` from a `QString`.
 - Support for further types: `QUuid`
 
 ### Fixed